Starting work to allow utc timestamps instead of numbers as migration versions

+++ Source/Machine.Migrations.Tests/Services/Impl/VersionStateFactoryTests.cs
@@ -34,11 +34,11 @@ namespace Machine.Migrations.Services.Impl
     {
       using (_mocks.Record())
       {
-        SetupResult.For(_configuration.DesiredVersion).Return((short)2);
-        SetupResult.For(_schemaStateManager.GetAppliedMigrationVersions(null)).Return(new short[] {1, 2, 3});
+        SetupResult.For(_configuration.DesiredVersion).Return((long)2);
+        SetupResult.For(_schemaStateManager.GetAppliedMigrationVersions(null)).Return(new long[] {1, 2, 3});
       }
       VersionState actual = _target.CreateVersionState(_migrations);
-      CollectionAssert.AreEqual(new short[] {1, 2, 3}, new List<short>(actual.Applied));
+      CollectionAssert.AreEqual(new long[] {1, 2, 3}, new List<long>(actual.Applied));
     }
 
     [Test]
@@ -46,8 +46,8 @@ namespace Machine.Migrations.Services.Impl
     {
       using (_mocks.Record())
       {
-        SetupResult.For(_configuration.DesiredVersion).Return((short)2);
-        SetupResult.For(_schemaStateManager.GetAppliedMigrationVersions(null)).Return(new short[] {1, 2, 3});
+        SetupResult.For(_configuration.DesiredVersion).Return((long)2);
+        SetupResult.For(_schemaStateManager.GetAppliedMigrationVersions(null)).Return(new long[] {1, 2, 3});
       }
       VersionState actual = _target.CreateVersionState(_migrations);
       Assert.AreEqual(4, actual.Last);
@@ -58,8 +58,8 @@ namespace Machine.Migrations.Services.Impl
     {
       using (_mocks.Record())
       {
-        SetupResult.For(_configuration.DesiredVersion).Return((short)-1);
-        SetupResult.For(_schemaStateManager.GetAppliedMigrationVersions(null)).Return(new short[] {1, 2, 3});
+        SetupResult.For(_configuration.DesiredVersion).Return((long)-1);
+        SetupResult.For(_schemaStateManager.GetAppliedMigrationVersions(null)).Return(new long[] {1, 2, 3});
       }
       VersionState actual = _target.CreateVersionState(_migrations);
       Assert.AreEqual(4, actual.Desired);
@@ -71,8 +71,8 @@ namespace Machine.Migrations.Services.Impl
     {
       using (_mocks.Record())
       {
-        SetupResult.For(_configuration.DesiredVersion).Return((short)5);
-        SetupResult.For(_schemaStateManager.GetAppliedMigrationVersions(null)).Return(new short[] {1, 2, 3});
+        SetupResult.For(_configuration.DesiredVersion).Return((long)5);
+        SetupResult.For(_schemaStateManager.GetAppliedMigrationVersions(null)).Return(new long[] {1, 2, 3});
       }
       _target.CreateVersionState(_migrations);
     }
@@ -83,7 +83,7 @@ namespace Machine.Migrations.Services.Impl
       using (_mocks.Record())
       {
         SetupResult.For(_configuration.Scope).Return("core");
-        SetupResult.For(_schemaStateManager.GetAppliedMigrationVersions("core")).Return(new short[] {1, 2, 3});
+        SetupResult.For(_schemaStateManager.GetAppliedMigrationVersions("core")).Return(new long[] {1, 2, 3});
       }
       VersionState actual = _target.CreateVersionState(_migrations);
       Assert.AreEqual(0, actual.Desired);
